+++ CHANGELOG.md
@@ -3,17 +3,22 @@
 # 0.0.5 (UNRELEASED)
 ## omnibus-marketplace
 * Update README.md
+* Use cloud-init to configure instances at boot time
+* 'preconfigure' marketplace images to make setup times shorter
 
 ## chef-marketplace-ctl
 * Add chef-marketplace-ctl register-node command
+* Add --preconfigure switch to chef-marketplace-ctl setup
 
 ## chef-marketplace-gem
 * Add Marketplace Api client
 * Add Node registration to setup
+* Update setup to support preconfiguring
 
 ## chef-marketplace-cookbooks
 * Add node registration recipe
 * Add marketplace_api config context to marketplace.rb
+* Add preconfigure recipes
 
 # 0.0.4 (2015-11-09)
 ## omnibus-marketplace